+++ test/functional-tests/javascript-obfuscator/fixtures/class-static-block-support-1.js
@@ -1,0 +1,11 @@
+let foo = 1;
+class Foo {
+    static abc;
+
+    static {
+        let bar = 2;
+        Foo.abc = 3;
+        foo = 4;
+        bar = 5;
+    }
+}